Project 1 Switches to Leds:
Created first VHDL project that connects the switches on the Go board (NandLand's proprietary board) to the LEDs on the Go board.

Created a constraint file labeling names to pins.

Created a .gitignore file to ignore all the crap that comes out of the build process so my git repository is just the files I create.