Project 4: Debouncing a Switch
This project correlated with Chapter 5 of the Getting Started with FPGAs. In this chapter, writing test benches were the focal point of the chapter.

It covered basic simulations with wave forms while also introducing the idea of self-checking testbenches.

Also added *.bak to gitignore since I found a file that was not needed on github with that extension.