Partial implementation of Project 5 which is associated to Chapter 6 of the book.
Had to rename the folder to just Project5 because it was leading to weird compilation issues.

This project uses either an LFSR or a counter to blink an LED. The LED is selected with a Mux, the selection lines are hooked up to buttons 1 and 2.

There may be more updates to this project later as I did not finish the chapter as of writing this.