Project 5: Reorganized the files to include two top files. One uses a counter to implement the blinking, another uses an LFSR.
Chapter 6 of the book also discusses RAM and gives an implementation that has been added to this project.

It is important to note that the RAM is not used to implement the project. It was just included in the same chapter as the project.

In Chapter 6, mux/demuxes, shift registers, RAM, and FIFO's were covered.